Thread

+++ java_project/Cp15-Thread/src/ex/Sum.java
@@ -2,17 +2,17 @@ package ex;
 
 public class Sum {
 
-	int num;
+	long num;
 	
 	Sum(){
 		num=0;
 	}
 	
-	void addNum(int n) {
+	void addNum(long n) {
 		num+=n; // num = num + n;
 	}
 	
-	int getNum() {
+	long getNum() {
 		return num;
 	}
 }
